+++ Tests/Functional/Backend/RequestForm/StartRequestFormTest.php
@@ -41,7 +41,7 @@ class StartRequestFormTest extends FunctionalTestCase
     /**
      * @var array
      */
-    protected $coreExtensionsToLoad = ['cms', 'core', 'frontend', 'version', 'lang', 'fluid', 'info'];
+    protected $coreExtensionsToLoad = ['cms', 'version', 'lang', 'info'];
 
     /**
      * @var StartRequestForm