+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 #### Classes
 #### Functions & Properties
 
+* Support for TYPO3 9
+
 ## Crawler 10.0.3
 Crawler 10.0.2 was released on September 29th, 2021