+++ CHANGELOG.md
@@ -20,6 +20,8 @@
 * FlushQueueCommand --page parameter is removed
 * Process->getTimeForFirstItem()
 * Process->getTimeForLastItem()
+* QueueRepository->countUnprocessedItems()
+* QueueRepository->countAllByProcessId()
 
 ## Crawler 10.0.3
 Crawler 10.0.2 was released on September 29th, 2021